demo/src/main/java/com/peper/blog/controllers/Authentication.java
package com.peper.blog.controllers;

import com.peper.blog.authentication.AuthenticationRequest;
import com.peper.blog.authentication.AuthenticationResponse;
import com.peper.blog.authentication.UserDetailsExtended;
import com.peper.blog.entities.PersonEntity;
import com.peper.blog.services.UserAuth;
import com.peper.blog.services.UserService;
import com.peper.blog.utilities.JwtUtils;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.security.authentication.AuthenticationManager;
import org.springframework.security.authentication.BadCredentialsException;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;


@RestController
@RequestMapping("/")
public class Authentication {

    @Autowired
    private UserAuth userAuth;

    @Autowired
    private AuthenticationManager authenticationManager;

    @Autowired
    private JwtUtils jwtUtils;

    @Autowired
    private UserService userService;

    @PostMapping(value = "/register", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE )
    public ResponseEntity createUser(@Valid @RequestBody PersonEntity personEntity){
        boolean success = userService.createPerson(personEntity);
        return ResponseEntity.status(HttpStatus.CREATED).build();
    }

    @PostMapping(value = "/login", produces = MediaType.APPLICATION_JSON_VALUE)
    public ResponseEntity<AuthenticationResponse> login(@Valid @RequestBody AuthenticationRequest authenticationRequest) throws Exception{
        try {
            authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(authenticationRequest.getEmail(),
                    authenticationRequest.getPassword()));
        }catch (BadCredentialsException e){
            throw new Exception("bad creds :",  e);
        }

        UserDetailsExtended userDetails = userAuth.loadUserByUsername(authenticationRequest.getEmail());

        String jwt = jwtUtils.generateToken(userDetails);

        return ResponseEntity.ok(new AuthenticationResponse(jwt));
    }

}
